+++ native/include/orvix/cli/capture_cli.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace orvix::cli {
+
+class CaptureCli final {
+public:
+    [[nodiscard]]
+    int run(int argc, char* argv[]) const;
+};
+
+}  // namespace orvix::cli